+++ ECommerce.WebApp/client/src/_configs/enviroment.config.ts
@@ -1,5 +1,6 @@
 export const ENV = {
+    CD_START: `cd D:data/dotnet/ecommerce-webapp/ECommerce.WebApp/client && npm start`,
     IMAGE_URL: 'https://hihichi.com/images',
     API_URL: 'https://hihichi.com/api',
-    // API_URL: 'https://localhost:44330/api'
+    // API_URL: 'https://localhost:44330/api',
 }
