+++ ECommerce.WebApp/client/src/_pages/home/_components/brand-item.component.tsx
@@ -3,7 +3,7 @@ import { ENV } from "src/_configs/enviroment.config";
 const BrandItem = (props: any) => {
     const { data } = props;
     return (
-        <div className="bran__list-item card" style={{ border: '.5px solid #ddd', padding: '12px' }}>
+        <div className="bran__list-item rounded" style={{ border: '.5px solid #ddd', padding: '12px' }}>
             <a href="/">
                 <div className="border-default bran__list-img flex items-center justify-between py-2 px-4 mx-auto bg-white">
                     <img src={ENV.IMAGE_URL + '/brand/' + data.brandImagePath} className="max-w-full max-h-full mx-auto" alt="name" />
