+++ ECommerce.WebApp/client/src/_pages/home/_components/highlight-brand-list.component.tsx
@@ -1,6 +1,4 @@
-import { useCallback, useEffect, useRef } from 'react';
-import { ICON_NAME } from 'src/_shares/_components/mui-icon/_enums/mui-icon.enum';
-import MuiIcon from 'src/_shares/_components/mui-icon/mui-icon.component';
+import { useRef } from 'react';
 import { Pagination, Navigation } from 'swiper';
 import { Swiper, SwiperSlide } from 'swiper/react';
 import { ENV } from 'src/_configs/enviroment.config';
@@ -8,19 +6,7 @@ import { useHomeStore } from 'src/_cores/_store/root-store';
 
 const HomeBrandList = () => {
   const homeStore = useHomeStore();
-  const navigationPrevRef = useRef<any>(null);
-  const navigationNextRef = useRef<any>(null);
   const sliderRef = useRef<any>(null);
-
-  const handlePrev = useCallback(() => {
-    if (!sliderRef.current) return;
-    sliderRef.current.swiper.slidePrev();
-  }, []);
-
-  const handleNext = useCallback(() => {
-    if (!sliderRef.current) return;
-    sliderRef.current.swiper.slideNext();
-  }, []);
 
   return (
     <div className='hidden md:block'>
@@ -37,16 +23,7 @@ const HomeBrandList = () => {
               ref={sliderRef}
               slidesPerView={5}
               spaceBetween={10}
-              navigation={{
-                enabled: true,
-                prevEl: navigationPrevRef.current,
-                nextEl: navigationNextRef.current,
-                disabledClass: 'pointer-events-none'
-              }}
-              onBeforeInit={(swiper) => {
-                swiper.navigation.nextEl = navigationNextRef.current;
-                swiper.navigation.prevEl = navigationPrevRef.current;
-              }}
+              navigation={true}
               modules={[Pagination, Navigation]}
               className='brand-slider my-4'
             >
@@ -58,18 +35,6 @@ const HomeBrandList = () => {
                 </SwiperSlide>
               ))}
             </Swiper>
-            <button className='absolute left-[-50px] bottom-1/2 translate-y-1/2' ref={navigationPrevRef} onClick={handlePrev}>
-              <MuiIcon
-                name={ICON_NAME.FEATHER.CHEVRON_LEFT}
-                className='w-[36px] h-[36px] border border-solid border-black hover:border-[#3b99fc] text-black hover:text-[#3b99fc]'
-              />
-            </button>
-            <button className='absolute right-[-50px] bottom-1/2 translate-y-1/2' ref={navigationNextRef} onClick={handleNext}>
-              <MuiIcon
-                name={ICON_NAME.FEATHER.CHEVRON_RIGHT}
-                className='w-[36px] h-[36px] border border-solid border-black hover:border-[#3b99fc] text-black hover:text-[#3b99fc]'
-              />
-            </button>
           </div>
           <div className="w-full text-center block">
             <a className="bran__viewmore inline-block" style={{ cursor: 'pointer' }} href="/">Xem thêm</a>
