+++ ECommerce.WebApp/client/src/_pages/home/_styles/banner.css
@@ -32,14 +32,23 @@
 
 .brand-slider .swiper-button-prev,
 .brand-slider .swiper-button-next {
-    width: 27px;
-    height: 27px;
-    border: 1px solid black;
-    border-radius: 50%;
-    opacity: .6;
+    color: black;
+    height: 100%;
+    top: 0;
+    margin-top: 0;
+}
+.brand-slider .swiper-button-prev {
+    left: 0;
+}
+.brand-slider .swiper-button-next {
+    right: 0;
 }
 .brand-slider .swiper-button-prev::after,
 .brand-slider .swiper-button-next::after {
-    font-size: 12px;
-    color: black;
+    font-size: 20px;
+}
+
+.brand-slider .swiper-button-prev:hover,
+.brand-slider .swiper-button-next:hover { 
+    background: rgba(0, 0, 0, 0.03);
 }
