+++ ECommerce.WebApp/client/src/_pages/product-detail/_components/layout/product-detail-footer.component.tsx
