+++ ECommerce.WebApp/client/src/_pages/product-detail/_components/silde/product-detail-real-images.component.tsx
@@ -2,7 +2,7 @@ import { Swiper, SwiperSlide } from "swiper/react";
 import "swiper/css";
 import "swiper/css/free-mode";
 import "swiper/css/pagination";
-import { FreeMode, Pagination } from "swiper";
+import { Autoplay, FreeMode, Pagination } from "swiper";
 
 const ProductDetailRealImages = () => {
   return (
@@ -18,7 +18,11 @@ const ProductDetailRealImages = () => {
         pagination={{
           clickable: true,
         }}
-        modules={[FreeMode, Pagination]}
+        autoplay={{
+          delay: 1500,
+          disableOnInteraction: false,
+        }}
+        modules={[Autoplay, FreeMode, Pagination]}
         className="mySwiper cursor-grab"
       >
         <SwiperSlide>
