+++ ECommerce.WebApp/client/src/_pages/product-detail/_interfaces/comment.interface.ts
@@ -1,0 +1,3 @@
+export interface IUserComment {
+    type: 'post' | 'comment' | 'reply'
+}
