+++ ECommerce.WebApp/client/src/_shares/_assets/_styles/main/product-detail.css
@@ -144,7 +144,7 @@
   }
 
   .product__detail-content .lSSlideWrapper {
-    height: 300px;
+    height: 375px;
   }
 }
