+++ ECommerce.WebApp/client/src/_shares/_components/mui-icon/_enums/mui-icon.enum.ts
@@ -15,6 +15,7 @@ export const ICON_NAME = {
     FACEBOOK: "FACEBOOK_ROUNDED",
     ARROW_LEFT: "ARROW_LEFT_ROUNDED",
     ARROW_RIGHT: "ARROW_RIGHT_ROUNDED",
+    STAR: "STAR_ROUNDED"
   },
   OUTLINED: {
     MAIL: "MAIL_OUTLINED",
@@ -44,6 +45,7 @@ export const ICON_NAME = {
     MENU: "MENU",
     X: "X",
     THUMBS_UP: 'THUMBS_UP',
-    THUMBS_DOWN: 'THUMBS_DOWN'
+    THUMBS_DOWN: 'THUMBS_DOWN',
+    PHONE_PRODUCT_DETAIL: 'PHONE_PRODUCT_DETAIL',
   },
 };
