+++ ECommerce.WebApp/client/src/_shares/_layouts/default-layout/banner.component.tsx
@@ -31,7 +31,6 @@ const Banner = () => {
     <div className={`${hideBanner ? 'hidden' : ''}`}>
       <Swiper
         className={`banner-slider`}
-        // autoplay={autoplay}
         loop={true}
         slidesPerView={1}
         navigation={true}
@@ -40,7 +39,6 @@ const Banner = () => {
         modules={[Autoplay, Pagination, Navigation]}
       >
         <SwiperSlide>
-          {/* <img className='h-auto my-auto' src="/assets/images/banner/2d83236a-01eb-4195-ab86-a4b12af64f79.jpg" alt="" /> */}
           <img
             className="w-full object-contain"
             src="https://hihichi.com/images/banner/banner_147f1f14-cf61-4808-b5cb-180a60d7fe8f.png"
@@ -48,7 +46,6 @@ const Banner = () => {
           />
         </SwiperSlide>
         <SwiperSlide>
-          {/* <img className='h-auto' src="/assets/images/banner/a9c4f5ca-0786-468d-b51c-68f9b856569b.jpg" alt="" /> */}
           <img
             className="w-full object-contain"
             src="https://hihichi.com/images/banner/banner_947667cf-bb07-4ab8-9248-9fd7428f7837.jpg"
